# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Order Flow API
  version: 1.0.0
  description: |
    Contract for the Order Flow HTTP API. Import this file directly into
    Postman, or attach it to Postman AI when generating collections, examples,
    mocks, tests, or server implementations.

    Operations marked `x-implementation-status: implemented` exist in the
    current API. Operations marked `planned` are contracts for upcoming work.
servers:
  - url: http://localhost:3001
    description: Local development
  - url: https://api.example.com
    description: Production placeholder
...
paths:
...
  /api/v1/orders:
    post:
      tags: [Service Orders]
      operationId: createOrder
      summary: Create a service order
      x-implementation-status: implemented
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          $ref: "#/components/responses/ValidationError"
    get:
      tags: [Service Orders]
      operationId: listOrders
      summary: List service orders
      x-implementation-status: implemented
      parameters:
        - $ref: "#/components/parameters/PageQuery"
        - $ref: "#/components/parameters/LimitQuery"
        - name: createdByUserId
          in: query
          schema:
            type: string
        - name: fulfillmentStatus
          in: query
          schema:
            $ref: "#/components/schemas/FulfillmentStatus"
        - name: paymentStatus
          in: query
          schema:
            $ref: "#/components/schemas/PaymentStatus"
        - name: assignedBaristaId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"
  /api/v1/orders/{orderId}:
    get:
      tags: [Service Orders]
      operationId: getOrder
      summary: Get order detail with timeline
      x-implementation-status: implemented
      parameters:
        - $ref: "#/components/parameters/OrderIdPath"
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/orders/{orderId}/cancel:
    post:
      tags: [Service Orders]
      operationId: cancelOrder
      summary: Cancel an unpaid order
      x-implementation-status: implemented
      parameters:
        - $ref: "#/components/parameters/OrderIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelOrderRequest"
      responses:
        "200":
          description: Order cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /api/v1/orders/{orderId}/claim:
    post:
      tags: [Barista]
      operationId: claimOrder
      summary: Atomically claim a queued order
      x-implementation-status: implemented
      parameters:
        - $ref: "#/components/parameters/OrderIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [baristaId]
              additionalProperties: false
              properties:
                baristaId:
                  type: string
      responses:
        "200":
          description: Claimed order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /api/v1/admin/reconciliations:
    get:
      tags: [Admin Finance]
      operationId: listReconciliations
      summary: List reconciliation review transactions
      x-implementation-status: implemented
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Reconciliation records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericDataListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/admin/reconciliations/{reconciliationId}:
    get:
      tags: [Admin Finance]
      operationId: getReconciliation
      summary: Get a reconciliation review transaction
      x-implementation-status: implemented
      security:
        - bearerAuth: []
      parameters:
        - name: reconciliationId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Reconciliation detail
          content:
            application/json:
              schema:
                type: object
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/admin/reconciliations/{reconciliationId}/resolve:
    post:
      tags: [Admin Finance]
      operationId: resolveReconciliation
      summary: Resolve a reconciliation review transaction
      x-implementation-status: implemented
      security:
        - bearerAuth: []
      parameters:
        - name: reconciliationId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResolveReconciliationRequest"
      responses:
        "200":
          description: Reconciliation resolved
          content:
            application/json:
              schema:
                type: object
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    OrderIdPath:
      name: orderId
      in: path
      required: true
      schema:
        type: string
...
    PageQuery:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitQuery:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: VALIDATION_ERROR
              message: Request validation failed
              details:
                - path: username
                  message: Required
    Unauthorized:
      description: Authentication failed or session expired
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: UNAUTHORIZED
              message: Invalid or expired token
    Forbidden:
      description: Authenticated account lacks permission or is inactive
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Requested resource was not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: The operation conflicts with existing data or business rules
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    PaymentStatus:
      type: string
      enum: [UNPAID, PENDING, PAID, UNDERPAID, OVERPAID, REVIEW]
    FulfillmentStatus:
      type: string
      enum: [PENDING_PAYMENT, QUEUED, PREPARING, READY, DELIVERED, CANCELLED]
    PaymentMethod:
      type: string
      enum: [QR, CASH]
...
    ResolutionAction:
      type: string
      enum: [NONE, LINK_MANUALLY, ACCEPT, REJECT, REFUND_REQUIRED]
    OrderStatusDomain:
      type: string
      enum: [PAYMENT, FULFILLMENT]
...
    GenericDataListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            type: object
...
    ResolveReconciliationRequest:
      type: object
      additionalProperties: false
      required: [resolutionAction, resolutionNote]
      properties:
        resolutionAction:
          $ref: "#/components/schemas/ResolutionAction"
        resolutionNote:
          type: string
          minLength: 1
...
    CreateOrderItemRequest:
      type: object
      additionalProperties: false
      required: [menuItemId, quantity]
      properties:
        menuItemId:
          type: string
        quantity:
          type: integer
          minimum: 1
        note:
          type: string
    CreateOrderRequest:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        paymentMethod:
          $ref: "#/components/schemas/PaymentMethod"
        customerNote:
          type: string
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/CreateOrderItemRequest"
    CancelOrderRequest:
      type: object
      additionalProperties: false
      required: [reason]
      properties:
        reason:
          type: string
          minLength: 1
...
    OrderItem:
      type: object
      required: [id, orderId, menuItemId, itemName, unitPrice, quantity]
      properties:
        id:
          type: string
        orderId:
          type: string
        menuItemId:
          type: string
        itemName:
          type: string
        unitPrice:
          type: string
          description: BigInt amount serialized by the runtime JSON layer.
        quantity:
          type: integer
        note:
          type: [string, "null"]
    OrderTimelineEntry:
      type: object
      required: [id, orderId, statusDomain, newStatus, createdAt]
      properties:
        id:
          type: string
        orderId:
          type: string
        statusDomain:
          $ref: "#/components/schemas/OrderStatusDomain"
        oldStatus:
          type: [string, "null"]
        newStatus:
          type: string
        changedByUserId:
          type: [string, "null"]
        reason:
          type: [string, "null"]
        createdAt:
          type: string
          format: date-time
    Order:
      type: object
      required: [id, orderCode, createdByUserId, paymentStatus, fulfillmentStatus, totalAmount, items, createdAt, updatedAt]
      properties:
        id:
          type: string
        orderCode:
          type: string
        createdByUserId:
          type: string
        assignedBaristaId:
          type: [string, "null"]
        paymentMethod:
          anyOf:
            - $ref: "#/components/schemas/PaymentMethod"
            - type: "null"
        paymentStatus:
          $ref: "#/components/schemas/PaymentStatus"
        fulfillmentStatus:
          $ref: "#/components/schemas/FulfillmentStatus"
        totalAmount:
          type: string
          description: BigInt amount serialized by the runtime JSON layer.
        customerNote:
          type: [string, "null"]
        cancellationReason:
          type: [string, "null"]
        paidAt:
          type: [string, "null"]
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        timeline:
          type: array
          items:
            $ref: "#/components/schemas/OrderTimelineEntry"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    OrderListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    ErrorDetail:
      type: object
      properties:
        path:
          type: string
        message:
          type: string
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            $ref: "#/components/schemas/ErrorDetail"
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorBody"